+++ test/smokes/code_sniffer/expectations.rb
@@ -2,7 +2,7 @@ require_relative "phpcs3/expectations"
 
 s = Runners::Testing::Smoke
 
-default_version = "3.5.7"
+default_version = "3.5.8"
 
 s.add_test(
   "success",