+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-[Full diff](https://github.com/sider/runners/compare/0.37.0...HEAD)
+[Full diff](https://github.com/sider/runners/compare/0.37.1...HEAD)
+
+## 0.37.1
+
+[Full diff](https://github.com/sider/runners/compare/0.37.0...0.37.1)
 
 - **devon_rex** 2.27.0 -> 2.27.1 [#1617](https://github.com/sider/runners/pull/1617)